+++ src/Block/Ssr/RecordList.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace Omikron\Factfinder\Block\Ssr;
 
+use Magento\Framework\App\Response\Http as Response;
+use Magento\Framework\App\Response\RedirectInterface;
 use Magento\Framework\Serialize\SerializerInterface;
 use Magento\Framework\View\Element\Template;
+use Omikron\Factfinder\Model\FieldRoles;
 use Omikron\Factfinder\Model\Ssr\SearchAdapter;
 
 class RecordList extends Template
@@ -18,15 +21,30 @@ class RecordList extends Template
     /** @var SerializerInterface */
     protected $jsonSerializer;
 
+    /** @var Response */
+    private $response;
+
+    /** @var RedirectInterface */
+    private $redirect;
+
+    /** @var FieldRoles */
+    private $fieldRoles;
+
     public function __construct(
-        Template\Context $context,
-        SearchAdapter $searchAdapter,
+        Template\Context    $context,
+        SearchAdapter       $searchAdapter,
         SerializerInterface $jsonSerializer,
-        array $data = []
+        Response            $response,
+        RedirectInterface   $redirect,
+        FieldRoles          $fieldRoles,
+        array               $data = []
     ) {
         parent::__construct($context, $data);
-        $this->searchAdapter       = $searchAdapter;
-        $this->jsonSerializer      = $jsonSerializer;
+        $this->searchAdapter = $searchAdapter;
+        $this->jsonSerializer = $jsonSerializer;
+        $this->redirect       = $redirect;
+        $this->response       = $response;
+        $this->fieldRoles     = $fieldRoles;
     }
 
     /**
@@ -41,7 +59,10 @@ class RecordList extends Template
             return "<ff-record-list ssr {$attributes}>";
         }, $html);
 
-        $result  = $this->searchResult($this->getRequest()->getParam('query', '*'), $this->getSearchParams());
+        $result = $this->searchResult($this->getRequest()->getParam('query', '*'), $this->getSearchParams());
+        if ($this->shouldRedirect($result)) {
+            $this->redirectToProductPage($result);
+        }
 
         // Add pre-rendered records
         $html = preg_replace_callback(self::RECORD_PATTERN, function (array $match) use ($result): string {
@@ -73,5 +94,17 @@ class RecordList extends Template
             [$key, $value] = array_map('urldecode', explode('=', $part));
             return $result + [$key => $value];
         }, []);
+    }
+
+    protected function redirectToProductPage(array $result): void
+    {
+        $this->redirect->redirect($this->response, $result['records'][0]['record'][$this->fieldRoles->getFieldRole('deeplink')]);
+    }
+
+    private function shouldRedirect(array $result): bool
+    {
+        $isExactSearch = isset($result['resultArticleNumberStatus']) && $result['resultArticleNumberStatus']
+            || isset($result['resultArticleNumberStatus']) && $result['resultArticleNumberStatus'] === 'resultsFound';
+        return count($result['records']) === 1 && $isExactSearch;
     }
 }