+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## Unreleased
 ### Added
  - Add possibility to select field to be exported as numerical in module configuration. Numerical fields will be exported in new multi-attribute column `NumericalAttributes`   
-
+ - Changing page is now update the page parameter in the canonical link (meta rel="canonical")  
+ 
 ## [v2.1.0] - 2021.07.15
 ### Added
  - Add possibility to export fields from variant level. Previously export took only configurable attributes from variants and rest was copied from parent