+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 ## Unreleased
 ### Added
- - Add possibility to select field to be exported as numerical in module configuration. Numerical fields will be exported in new multi-attribute column `NumericalAttributes`   
+  - SSR
+    - Added Single Hit Redirection feature when SSR is enabled
+  - Export
+    - Add possibility to select field to be exported as numerical in module configuration. Numerical fields will be exported in new multi-attribute column `NumericalAttributes`   
 
 ## [v2.1.0] - 2021.07.15
 ### Added