+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   - Export
     - Add possibility to select field to be exported as numerical in module configuration. Numerical fields will be exported in new multi-attribute column `NumericalAttributes`   
 
+### Fix
+ - SSR
+    - after first search with 0 result, next will render no products even if FACT-Finder returns them
+
 ## [v2.1.0] - 2021.07.15
 ### Added
  - Add possibility to export fields from variant level. Previously export took only configurable attributes from variants and rest was copied from parent